+++ blocksatcli/api/listen.py
@@ -5,8 +5,8 @@ import subprocess
 
 from . import net
 from . import msg as api_msg
-from .order import ApiOrder
-from .pkt import BlocksatPkt, BlocksatPktHandler, ApiChannel
+from .order import ApiOrder, ApiChannel
+from .pkt import BlocksatPkt, BlocksatPktHandler
 
 logger = logging.getLogger(__name__)